requirements.txt
Flask==2.1.2
flask_restx~=0.5.1
Werkzeug==2.1.2
gevent~=21.12.0
pg8000==1.29.1
PyJWT==2.4.0
cryptography~=37.0.2